ci(cypress): add linting to test files
