+++ .github/workflows/nodejs.yaml
@@ -26,6 +26,8 @@ jobs:
     - run: npm ci
     - run: npm run lint
       if: ${{ matrix.node-version == '16.x' }}
+    - run: npm run lint:cypress
+      if: ${{ matrix.node-version == '16.x' }}
     - run: npm run test-ci
     - uses: codecov/codecov-action@v1.2.1
       name: Upload Coverage Results
